+++ src/Struct/Marketplace/CategoryInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Scn\EvalancheSoapStruct\Struct\Marketplace;
+
+use Scn\EvalancheSoapStruct\Struct\StructInterface;
+
+interface CategoryInterface extends StructInterface
+{
+    public function getId(): int;
+
+    public function getText(): array;
+}